refactor: FSD 레이어 구조 변경
