+++ apps/web/src/entities/feed/index.ts
@@ -1,6 +1,1 @@
 export type { FeedPage, FeedPost, FeedTag } from './model/types';
-export {
-  FEED_QUERY_KEY,
-  useInfiniteFeedQuery,
-} from './model/useInfiniteFeedQuery';
-export { default as FeedCard } from './ui/FeedCard';
